Joint model for BIG data using FastJM

function for joint model in BIG DATA using FastJM

jmcsBig(dtlong, dtsurv, longm, survm, samplesize = 50, rd, id)

Arguments

dtlong

longitudinal dataset, which contains id,visit time,longitudinal measurements along with various covariates

dtsurv

survival dataset corresponding to the longitudinal dataset, with survival status and survival time

longm

model for longitudinal response

survm

survival model

samplesize

sample size to divide the Big data

rd

random effect part

id

name of id column in longitudinal dataset

Value

returns a list containing various output which are useful for prediction.

References

Li, Shanpeng, et al. "Efficient Algorithms and Implementation of a Semiparametric Joint Model for Longitudinal and Competing Risk Data: With Applications to Massive Biobank Data." Computational and Mathematical Methods in Medicine 2022 (2022).

See also

Author

Atanu Bhattacharjee, Bhrigu Kumar Rajbongshi and Gajendra Kumar Vishwakarma

Examples

  # \donttest{
##
library(survival)
library(dplyr)
fit2<-jmcsBig(dtlong=data.frame(long2),dtsurv = data.frame(surv2),
longm=y~ x7+visit,survm=Surv(time,status)~x1+visit,rd= ~ visit|id,samplesize=200,id='id')
#> Error in jmcs(long.formula = longm, ydata = data.frame(dtlong1[[i]]),     surv.formula = survm, cdata = data.frame(dtsurv1[[i]]), random = rd): could not find function "jmcs"
P2<-survfitJMCS(model<-fit2,ids<-c(400),estimator = 'median')
#> Error in survfitJMCS(model <- fit2, ids <- c(400), estimator = "median"): unused argument (estimator = "median")
plot(P2$p1[[1]],estimator="both",include.y = TRUE)
#> Error in eval(expr, envir, enclos): object 'P2' not found
##
  # }