print.joinRMLBig

print.joinRMLBig

# S3 method for joinRMLBig
print(object, digits = 4, ...)

Arguments

object

object

digits

used for round the numeric values after decimal

...

others

Value

prints table containing various parameter estimates, SE, P- value for both survival and longitudinal submodel, if the model is bayesian it includes their credible interval too.

Examples

 # \donttest{
##
library(survival)
library(dplyr)
mod4<-joinRMLBig(dtlong=long2,
dtsurv = surv2,
longm=y~ x7+visit,
survm=Surv(time,status)~x1+visit,
rd=~ visit|id,
timeVar='visit',
samplesize=200,
id='id')
#> EM algorithm has converged!
#> Calculating post model fit statistics...
#> EM algorithm has converged!
#> Calculating post model fit statistics...
#> EM algorithm has converged!
#> Calculating post model fit statistics...
#> EM algorithm has converged!
#> Calculating post model fit statistics...
#> EM algorithm has converged!
#> Calculating post model fit statistics...
print(mod4)
#> 
#>  Joint model for Big data using joineRML
#>  Call: 
#> joinRMLBig(dtlong = long2, dtsurv = surv2, longm = y ~ x7 + visit, 
#>     survm = Surv(time, status) ~ x1 + visit, samplesize = 200, 
#>     rd = ~visit | id, timeVar = "visit", id = "id")
#> 
#> 
#>  Total observation in longitudinal data: 1000 
#> 
#>  Chunk size: 200 
#> 
#>  Longitudinal process: 
#>               Estimate    SE   Zvalue Pvalue
#> (Intercept)_1    8.771 0.175   50.130      0
#> x7_1            -0.022 0.000 -536.692      0
#> visit_1         -0.085 0.002  -35.703      0
#> sigma2_1         0.599 0.000 3174.213      0
#> 
#>  Survival process: 
#>          Estimate    SE  ZValue Pvalue
#> x11       -0.012 0.060  -0.205  0.837
#> visit     -0.140 0.007 -20.201  0.000
#> gamma_1   -0.011 0.005  -2.126  0.034
#> 
#>  Variance Covariance matrix of Random effects:
#>           Intercept  visit
#> Intercept     1.975 -0.352
#> visit        -0.352  0.132
   # }