Prediction using FastJM

prediction of survival probability using FastJM for BIG data

Usage

survfitJMCS(model, ids, u, method = "GH", obs.time)

Arguments

model

fitted model object

ids

value of id

u

see survfitjmcs

method

options are 'Laplace','GH'

obs.time

vector which represents time variable in the longitudinal data

Value

list of predicted value for the given id along with other information relevant for survival probability confidence plot

Examples

  # \donttest{
##
library(survival)
library(dplyr)
jmcs1<-jmcsBig(dtlong=data.frame(long2),
dtsurv = data.frame(surv2),
longm=y~ x7+visit,
survm=Surv(time,status)~x1+visit,
rd= ~ visit|id,
samplesize=200,id='id')
mod2<-jmcs1
P2<-survfitJMCS(model<-mod2,ids<-c(5),u<-seq(surv2[surv2$id==5,]$time,
surv2[surv2$id==5,]$time+10,0.2),obs.time='time')
print(P2)
#> 
#> Prediction of Conditional Probabilities of Event
#> based on the pseudo-adaptive Guass-Hermite quadrature rule with 6 quadrature points
#> $`5`
#>        times  PredSurv
#> 1   2.351878 1.0000000
#> 2   2.351878 1.0000000
#> 3   2.551878 0.9893133
#> 4   2.751878 0.9573583
#> 5   2.951878 0.9573583
#> 6   3.151878 0.9573583
#> 7   3.351878 0.9573583
#> 8   3.551878 0.9573583
#> 9   3.751878 0.9573583
#> 10  3.951878 0.9464924
#> 11  4.151878 0.9356432
#> 12  4.351878 0.9356432
#> 13  4.551878 0.9139051
#> 14  4.751878 0.8809480
#> 15  4.951878 0.8809480
#> 16  5.151878 0.8589332
#> 17  5.351878 0.8368443
#> 18  5.551878 0.8145375
#> 19  5.751878 0.7922527
#> 20  5.951878 0.7922527
#> 21  6.151878 0.7809792
#> 22  6.351878 0.7584506
#> 23  6.551878 0.7361916
#> 24  6.751878 0.7250806
#> 25  6.951878 0.7137624
#> 26  7.151878 0.7023804
#> 27  7.351878 0.6910631
#> 28  7.551878 0.6573517
#> 29  7.751878 0.6573517
#> 30  7.951878 0.6461274
#> 31  8.151878 0.6123296
#> 32  8.351878 0.5893782
#> 33  8.551878 0.5777416
#> 34  8.751878 0.5548858
#> 35  8.951878 0.5548858
#> 36  9.151878 0.5436819
#> 37  9.351878 0.5217708
#> 38  9.551878 0.5108233
#> 39  9.751878 0.4998672
#> 40  9.951878 0.4670233
#> 41 10.151878 0.4563563
#> 42 10.351878 0.4449410
#> 43 10.551878 0.4449410
#> 44 10.751878 0.4331449
#> 45 10.951878 0.4213851
#> 46 11.151878 0.4093684
#> 47 11.351878 0.4093684
#> 48 11.551878 0.4093684
#> 49 11.751878 0.4093684
#> 50 11.951878 0.4093684
#> 51 12.151878 0.4093684
#> 52 12.351878 0.4093684
#> 
##
  # }